Code structure: make it better Fixes #10
Splitted in folders and single component files:
 - Components
 - Widgets

Added docstrings to all
