+++ Components/Widgets/InputPopup.js
@@ -1,0 +1,271 @@
+import { EventEmitter } from "events"
+import { ConsoleManager } from "../../index.js"
+
+/**
+ * @class InputPopup
+ * @extends EventEmitter
+ * @description This class is used to create a popup with a text or numeric input. 
+ * Emits the following events: 
+ * - "confirm" when the user confirm the input
+ * - "cancel" when the user cancel the input
+ * - "exit" when the user exit the input
+ * @param {string} id - The id of the popup.
+ * @param {string} title - The title of the popup.
+ * @param {string | number} value - The value of the input.
+ * @param {boolean} numeric - If the input is numeric.
+ * @param {boolean} visible - If the popup is visible. Default is false (make it appears using show()).
+ * 
+ * @example const popup = new InputPopup("popup1", "Choose the number", selectedNumber, true).show().on("confirm", (value) => { console.log(value) }) // show the popup and wait for the user to confirm
+ */
+export class InputPopup extends EventEmitter {
+    constructor(id, title, value, numeric, visible = false) {
+        super()
+            /** @const {ConsoleManager} CM the instance of ConsoleManager (singleton) */
+        this.CM = new ConsoleManager()
+        this.id = id
+        this.title = title
+        this.value = value
+        this.numeric = numeric
+        this.visible = visible
+        this.marginTop = 4
+        if (this.CM.widgetsCollection[this.id]) {
+            this.CM.unRegisterWidget(this)
+            const message = `InputPopup ${this.id} already exists.`
+            this.CM.error(message)
+            throw new Error(message)
+        }
+        this.CM.registerWiget(this)
+    }
+
+    /**
+     * @function keyListnerNumeric(str, key)
+     * @description This function is used to make the ConsoleManager handle the key events when the input is numeric and it is showed.
+     * Inside this function are defined all the keys that can be pressed and the actions to do when they are pressed.
+     * @param {string} str - The string of the input.
+     * @param {Object} key - The key object.
+     * @memberof InputPopup
+     */
+    keyListnerNumeric(str, key) {
+        const v = Number(this.value)
+        if (Number.isNaN(v)) {
+            v = 0
+        }
+        if (!Number.isNaN(Number(key.name))) {
+            if (v.toString().length < 20) {
+                let tmp = v.toString()
+                tmp += key.name
+                this.value = Number(tmp)
+            }
+            // To change the sign I check for the keys "+" and "-"
+        } else if (key.sequence === '-') {
+            this.value = v * -1
+        } else if (key.sequence === '+') {
+            this.value = Math.abs(v)
+        } else {
+            switch (key.name) {
+                case 'backspace':
+                    // If backspace is pressed I remove the last character from the typed value
+                    if (v.toString().length > 0) {
+                        this.value = Number(v.toString().slice(0, v.toString().length - 1))
+                    }
+                    break
+                case 'return':
+                    {
+                        this.emit(`confirm`, this.value)
+                        this.CM.unRegisterWidget(this)
+                        this.hide()
+                        delete this
+                    }
+                    break
+                case 'escape':
+                    {
+                        this.emit(`cancel`)
+                        this.CM.unRegisterWidget(this)
+                        this.hide()
+                        delete this
+                    }
+                    break
+                case 'q':
+                    {
+                        this.CM.emit('exit')
+                        this.CM.unRegisterWidget(this)
+                        this.hide()
+                        delete this
+                    }
+                    break
+                default:
+                    break
+            }
+        }
+        this.CM.refresh()
+    }
+
+    /**
+     * @function keyListnerText(str, key)
+     * @description This function is used to make the ConsoleManager handle the key events when the input is text and it is showed.
+     * Inside this function are defined all the keys that can be pressed and the actions to do when they are pressed.
+     * @param {string} str - The string of the input.
+     * @param {Object} key - The key object.
+     * @memberof InputPopup
+     */
+    keyListnerText(str, key) {
+        const v = this.value
+        if (v.toString().length < 20) {
+            let tmp = v.toString()
+            tmp += key.name
+            this.value = tmp
+        }
+        switch (key.name) {
+            case 'backspace':
+                // If backspace is pressed I remove the last character from the typed value
+                if (v.toString().length > 0) {
+                    this.value = v.toString().slice(0, v.toString().length - 1)
+                }
+                break
+            case 'return':
+                {
+                    this.emit(`confirm`, this.value)
+                    this.CM.unRegisterWidget(this)
+                    this.hide()
+                    delete this
+                }
+                break
+            case 'escape':
+                {
+                    this.emit(`cancel`)
+                    this.CM.unRegisterWidget(this)
+                    this.hide()
+                    delete this
+                }
+                break
+            default:
+                break
+        }
+        this.CM.refresh()
+    }
+
+    /**
+     * @description This function is used to get the value of the input.
+     * @returns {string | number} The value of the input.
+     * @memberof InputPopup
+     */
+    getValue() {
+        return this.value
+    }
+
+    /**
+     * @description This function is used to change the value of the input. It also refresh the ConsoleManager.
+     * @param {string | number} newValue - The new value of the input.
+     * @memberof InputPopup
+     * @returns {InputPopup} The instance of the InputPopup.
+     */
+    setValue(newValue) {
+        this.value = newValue
+        this.CM.refresh()
+        return this
+    }
+
+    /**
+     * @description This function is used to show the popup. It also register the key events and refresh the ConsoleManager.
+     * @returns {InputPopup} The instance of the InputPopup.
+     * @memberof InputPopup
+     */
+    show() {
+        if (!this.visible) {
+            this.manageInput()
+            this.visible = true
+            this.CM.refresh()
+        }
+        return this
+    }
+
+    /**
+     * @description This function is used to hide the popup. It also unregister the key events and refresh the ConsoleManager.
+     * @returns {InputPopup} The instance of the InputPopup.
+     * @memberof InputPopup
+     */
+    hide() {
+        if (this.visible) {
+            this.unManageInput()
+            this.visible = false
+            this.CM.refresh()
+        }
+        return this
+    }
+
+    /**
+     * @description This function is used to get the visibility of the popup.
+     * @returns {boolean} The visibility of the popup.
+     * @memberof InputPopup
+     */
+    isVisible() {
+        return this.visible
+    }
+
+    /**
+     * @description This function is used to add the InputPopup key listener callback to te ConsoleManager.
+     * @returns {InputPopup} The instance of the InputPopup.
+     * @memberof InputPopup
+     */
+    manageInput() {
+        // Add a command input listener to change mode
+        if (this.numeric) {
+            this.CM.setKeyListener(this.id, this.keyListnerNumeric.bind(this))
+        } else {
+            this.CM.setKeyListener(this.id)
+        }
+        return this
+    }
+
+    /**
+     * @description This function is used to remove the InputPopup key listener callback to te ConsoleManager.
+     * @returns {InputPopup} The instance of the InputPopup.
+     * @memberof InputPopup
+     */
+    unManageInput() {
+        // Add a command input listener to change mode
+        if (this.numeric) {
+            this.CM.removeKeyListener(this.id, this.keyListnerNumeric.bind(this))
+        } else {
+            this.CM.removeKeyListener(this.id)
+        }
+        return this
+    }
+
+    /**
+     * @description This function is used to draw the InputPopup to the screen in the middle.
+     * @returns {InputPopup} The instance of the InputPopup.
+     * @memberof InputPopup
+     */
+    draw() {
+        const offset = 2
+        const windowWidth = this.title.length > this.value.toString().length ? this.title.length + (2 * offset) : this.value.toString().length + (2 * offset) + 1
+        const halfWidth = Math.round((windowWidth - this.title.length) / 2)
+        let header = "┌"
+        for (let i = 0; i < windowWidth; i++) {
+            header += "─"
+        }
+        header += "┐\n"
+        header += `│${" ".repeat(halfWidth)}${this.title}${" ".repeat(windowWidth - halfWidth - this.title.length)}│\n`
+        header += "├" + "─".repeat(windowWidth) + "┤\n"
+
+        let footer = "└"
+        for (let i = 0; i < windowWidth; i++) {
+            footer += "─"
+        }
+        footer += "┘\n"
+
+        let content = ""
+            // Draw an input field
+        content += `│${"> "}${this.value}█${" ".repeat(windowWidth - this.value.toString().length - 3)}│\n`
+
+        const windowDesign = `${header}${content}${footer}`
+        windowDesign.split('\n').forEach((line, index) => {
+            this.CM.Screen.cursorTo(Math.round((this.CM.Screen.width / 2) - (windowWidth / 2)), this.marginTop + index)
+            this.CM.Screen.write({ text: line, style: { color: "white" } })
+        })
+        return this
+    }
+}
+
+export default InputPopup
